Modularize patching logic
