+++ pkg/common/tag.go
@@ -1,0 +1,35 @@
+package common
+
+import (
+	"fmt"
+
+	"github.com/distribution/reference"
+)
+
+// ResolvePatchedTag merges explicit tag & suffix rules, returning the final patched tag.
+func ResolvePatchedTag(imageRef reference.Named, explicitTag, suffix string) (string, error) {
+	// if user explicitly sets a final tag, that wins outright
+	if explicitTag != "" {
+		return explicitTag, nil
+	}
+
+	// parse out any existing tag from the image ref
+	var baseTag string
+	if tagged, ok := imageRef.(reference.Tagged); ok {
+		baseTag = tagged.Tag()
+	}
+
+	// if suffix is empty, default to "patched"
+	if suffix == "" {
+		suffix = "patched"
+	}
+
+	// if we have no original baseTag (the user's image had no tag),
+	// then we can't append a suffix to it
+	if baseTag == "" {
+		return "", fmt.Errorf("no tag found in image reference %s", imageRef.String())
+	}
+
+	// otherwise, combine them
+	return fmt.Sprintf("%s-%s", baseTag, suffix), nil
+}
